media-source/SourceBuffer-appendStreamWithMaxSize_readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#appendStream(stream, maxSize) when readyState is in 'ended' state</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'stream';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(xhr.response);
    };
}
function GET2(url, processBody) {
    GET(FILE, function(stream1) {
        GET(FILE, function(stream2) {
            processBody(stream1, stream2);
        });
    });
}
//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    GET2(FILE, function(stream1, stream2) {
        var mediaSource = new MediaSource();
        var sourceopenNum = 0;
        mediaSource.addEventListener('sourceopen', function(e){
            sourceopenNum++;
            if(sourceopenNum > 1) {
                t.done();
                return;
            }
            this.sourceBuffer = mediaSource.addSourceBuffer(mime);
            this.sourceBuffer.addEventListener('updateend', function() {
                mediaSource.endOfStream();
            });
            this.sourceBuffer.appendStream(stream1, 1024);
        });
        mediaSource.addEventListener('sourceended', function(e){
            if(sourceopenNum > 1) {
                return;
            }
            t.step(function() {
                assert_equals(this.readyState, 'ended');
                this.sourceBuffer.appendStream(stream2, 1024);//append new video stream
                assert_equals(this.readyState, 'open');
            }.bind(this));
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#appendStream(stream, maxSize) : If the readyState attribute is in the "ended", should run the following steps:' +
'1. Set the readyState attribute of the parent media source to "open" ' +
'2. Queue a task to fire a simple event named sourceopen at the parent media source');

</script>
</body>
</html>
